+++ examples/next-openai/agent/openai-web-search-agent.ts
@@ -1,7 +1,7 @@
-import { openai } from '@ai-sdk/openai';
-import { Agent, InferAgentUIMessage } from 'ai';
+import { openai, OpenAIResponsesProviderOptions } from '@ai-sdk/openai';
+import { BasicAgent, InferAgentUIMessage } from 'ai';
 
-export const openaiWebSearchAgent = new Agent({
+export const openaiWebSearchAgent = new BasicAgent({
   model: openai('gpt-5-mini'),
   tools: {
     web_search: openai.tools.webSearch({
@@ -13,6 +13,13 @@ export const openaiWebSearchAgent = new Agent({
         country: 'US',
       },
     }),
+  },
+  providerOptions: {
+    openai: {
+      reasoningEffort: 'medium',
+      reasoningSummary: 'detailed',
+      // store: false,
+    } satisfies OpenAIResponsesProviderOptions,
   },
   onStepFinish: ({ request }) => {
     console.dir(request.body, { depth: Infinity });