+++ examples/next-openai/app/use-chat-persistence-metadata/[chatId]/chat.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import ChatInput from '@/component/chat-input';
+import ChatInput from '@/components/chat-input';
 import { zodSchema } from '@ai-sdk/provider-utils';
 import { UIMessage, useChat } from '@ai-sdk/react';
 import { DefaultChatTransport } from 'ai';