+++ packages/openai/src/responses/convert-to-openai-responses-input.ts
@@ -6,6 +6,7 @@ import {
 } from '@ai-sdk/provider';
 import {
   convertToBase64,
+  isNonNullable,
   parseProviderOptions,
   validateTypes,
 } from '@ai-sdk/provider-utils';
@@ -141,18 +142,38 @@ export async function convertToOpenAIResponsesInput({
         for (const part of content) {
           switch (part.type) {
             case 'text': {
+              const id = part.providerOptions?.openai?.itemId as
+                | string
+                | undefined;
+
+              // item references reduce the payload size
+              if (store && id != null) {
+                input.push({ type: 'item_reference', id });
+                break;
+              }
+
               input.push({
                 role: 'assistant',
                 content: [{ type: 'output_text', text: part.text }],
-                id:
-                  (part.providerOptions?.openai?.itemId as string) ?? undefined,
+                id,
               });
+
               break;
             }
             case 'tool-call': {
               toolCallParts[part.toolCallId] = part;
 
               if (part.providerExecuted) {
+                break;
+              }
+
+              const id = part.providerOptions?.openai?.itemId as
+                | string
+                | undefined;
+
+              // item references reduce the payload size
+              if (store && id != null) {
+                input.push({ type: 'item_reference', id });
                 break;
               }
 
@@ -164,9 +185,7 @@ export async function convertToOpenAIResponsesInput({
                 input.push({
                   type: 'local_shell_call',
                   call_id: part.toolCallId,
-                  id:
-                    (part.providerOptions?.openai?.itemId as string) ??
-                    undefined,
+                  id: id!,
                   action: {
                     type: 'exec',
                     command: parsedInput.action.command,
@@ -185,8 +204,7 @@ export async function convertToOpenAIResponsesInput({
                 call_id: part.toolCallId,
                 name: part.toolName,
                 arguments: JSON.stringify(part.input),
-                id:
-                  (part.providerOptions?.openai?.itemId as string) ?? undefined,
+                id,
               });
               break;
             }
@@ -219,8 +237,9 @@ export async function convertToOpenAIResponsesInput({
                 const reasoningMessage = reasoningMessages[reasoningId];
 
                 if (store) {
+                  // use item references to refer to reasoning (single reference)
+                  // when the first part is encountered
                   if (reasoningMessage === undefined) {
-                    // use item references to refer to reasoning (single reference)
                     input.push({ type: 'item_reference', id: reasoningId });
 
                     // store unused reasoning message to mark id as used
@@ -259,6 +278,12 @@ export async function convertToOpenAIResponsesInput({
                     input.push(reasoningMessages[reasoningId]);
                   } else {
                     reasoningMessage.summary.push(...summaryParts);
+
+                    // updated encrypted content to enable setting it in the last summary part:
+                    if (providerOptions?.reasoningEncryptedContent != null) {
+                      reasoningMessage.encrypted_content =
+                        providerOptions.reasoningEncryptedContent;
+                    }
                   }
                 }
               } else {
@@ -311,25 +336,38 @@ export async function convertToOpenAIResponsesInput({
               contentValue = JSON.stringify(output.value);
               break;
             case 'content':
-              contentValue = output.value.map(item => {
-                switch (item.type) {
-                  case 'text': {
-                    return { type: 'input_text' as const, text: item.text };
+              contentValue = output.value
+                .map(item => {
+                  switch (item.type) {
+                    case 'text': {
+                      return { type: 'input_text' as const, text: item.text };
+                    }
+
+                    case 'image-data': {
+                      return {
+                        type: 'input_image' as const,
+                        image_url: `data:${item.mediaType};base64,${item.data}`,
+                      };
+                    }
+
+                    case 'file-data': {
+                      return {
+                        type: 'input_file' as const,
+                        filename: item.filename ?? 'data',
+                        file_data: `data:${item.mediaType};base64,${item.data}`,
+                      };
+                    }
+
+                    default: {
+                      warnings.push({
+                        type: 'other',
+                        message: `unsupported tool content part type: ${item.type}`,
+                      });
+                      return undefined;
+                    }
                   }
-                  case 'media': {
-                    return item.mediaType.startsWith('image/')
-                      ? {
-                          type: 'input_image' as const,
-                          image_url: `data:${item.mediaType};base64,${item.data}`,
-                        }
-                      : {
-                          type: 'input_file' as const,
-                          filename: 'data',
-                          file_data: `data:${item.mediaType};base64,${item.data}`,
-                        };
-                  }
-                }
-              });
+                })
+                .filter(isNonNullable);
               break;
           }