+++ packages/google-vertex/CHANGELOG.md
@@ -1,5 +1,92 @@
 # @ai-sdk/google-vertex
 
+## 4.0.0-beta.39
+
+### Patch Changes
+
+- Updated dependencies [ca07285]
+  - @ai-sdk/anthropic@3.0.0-beta.31
+
+## 4.0.0-beta.38
+
+### Patch Changes
+
+- Updated dependencies [9354297]
+  - @ai-sdk/anthropic@3.0.0-beta.30
+
+## 4.0.0-beta.37
+
+### Patch Changes
+
+- Updated dependencies [3794514]
+  - @ai-sdk/provider-utils@4.0.0-beta.19
+  - @ai-sdk/anthropic@3.0.0-beta.29
+  - @ai-sdk/provider@3.0.0-beta.8
+  - @ai-sdk/google@3.0.0-beta.26
+
+## 4.0.0-beta.36
+
+### Patch Changes
+
+- Updated dependencies [81d4308]
+- Updated dependencies [81d4308]
+  - @ai-sdk/anthropic@3.0.0-beta.28
+  - @ai-sdk/provider@3.0.0-beta.7
+  - @ai-sdk/google@3.0.0-beta.25
+  - @ai-sdk/provider-utils@4.0.0-beta.18
+
+## 4.0.0-beta.35
+
+### Patch Changes
+
+- Updated dependencies [4c5a6be]
+  - @ai-sdk/anthropic@3.0.0-beta.27
+
+## 4.0.0-beta.34
+
+### Patch Changes
+
+- Updated dependencies [f33a018]
+  - @ai-sdk/anthropic@3.0.0-beta.26
+
+## 4.0.0-beta.33
+
+### Patch Changes
+
+- Updated dependencies [703459a]
+  - @ai-sdk/provider-utils@4.0.0-beta.17
+  - @ai-sdk/anthropic@3.0.0-beta.25
+  - @ai-sdk/google@3.0.0-beta.24
+
+## 4.0.0-beta.32
+
+### Patch Changes
+
+- Updated dependencies [d08308b]
+  - @ai-sdk/anthropic@3.0.0-beta.24
+
+## 4.0.0-beta.31
+
+### Patch Changes
+
+- Updated dependencies [6f845b4]
+  - @ai-sdk/anthropic@3.0.0-beta.23
+
+## 4.0.0-beta.30
+
+### Patch Changes
+
+- f8c981f: Fix adding google search along with url context in vertex ai
+- Updated dependencies [f8c981f]
+  - @ai-sdk/google@3.0.0-beta.23
+
+## 4.0.0-beta.29
+
+### Patch Changes
+
+- Updated dependencies [09ba2dd]
+  - @ai-sdk/google@3.0.0-beta.22
+
 ## 4.0.0-beta.28
 
 ### Patch Changes