+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,102 @@
 
 ### Patch Changes
 
+- Updated dependencies [3794514]
+  - ai@6.0.0-beta.56
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [42cf7ed]
+  - ai@6.0.0-beta.55
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [9388ff1]
+  - ai@6.0.0-beta.54
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.53
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.52
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [5e313e3]
+  - ai@6.0.0-beta.51
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [4812235]
+- Updated dependencies [81d4308]
+  - ai@6.0.0-beta.50
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [703459a]
+  - ai@6.0.0-beta.49
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [7f2c9b6]
+  - ai@6.0.0-beta.48
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [c62ecf0]
+  - ai@6.0.0-beta.47
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- ai@6.0.0-beta.46
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [48454ab]
+  - ai@6.0.0-beta.45
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [2b1bf9d]
+  - ai@6.0.0-beta.44
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
+- Updated dependencies [27e8c3a]
+  - ai@6.0.0-beta.43
+
+## 0.0.1-beta.0
+
+### Patch Changes
+
 - ai@6.0.0-beta.42
 
 ## 0.0.1-beta.0